+++ examples/vite/src/App.jsx
@@ -1,8 +1,9 @@
+import cn from 'clsx';
 import { sv } from 'stilvoll';
 
 function App() {
   return (
-    <div className={sv.flex.items_center.gap_l}>
+    <div className={cn(sv.flex, sv.items_center, sv.gap_l)}>
       <div className={sv.invisible}>
         OMG, typesafe utility classes, so nice.
       </div>