+++ packages/core/src/index.ts
@@ -4,8 +4,7 @@ export {
   mergeWithDefaultConfig,
   loadUserConfig,
   createConfig,
-} from './lib/options';   
+} from './lib/options';
 export { extractClassNamesFromString } from './lib/extract';
-export { hashClassName } from './lib/utils';
 export { defaultFileSystemGlobs, defaultTypeDefinitionsPath } from './defaults';
 export * from './types';