+++ packages/core/src/lib/parser.ts
@@ -206,7 +206,7 @@ ${commentCode}
 "${selector}": Property;`;
   });
 
-  return `type Property = UtilityMap & string;
+  return `type Property = string;
 
 type UtilityMap = {
   ${properties.join('\n  ')}
@@ -312,12 +312,12 @@ const parseInputCSS = (
 
   const breakpoints: Record<string, string> = hasBreakpointsDefined
     ? Object.entries(options.breakpoints).reduce(
-      (acc, [key, value]) => ({
-        ...acc,
-        [key]: `@media screen and (min-width: ${value}px)`,
-      }),
-      {},
-    )
+        (acc, [key, value]) => ({
+          ...acc,
+          [key]: `@media screen and (min-width: ${value}px)`,
+        }),
+        {},
+      )
     : foundBreakpoints;
 
   return {