+++ packages/stilvoll/src/index.ts
@@ -1,2 +1,1 @@
-export { cn } from './lib/classnames';
 export { sv } from './lib/stilvoll';